Add header to avalanche forecast tab
