+++ components/forecast/AvalancheForecast.tsx
@@ -1,29 +1,19 @@
 import React, {useEffect} from 'react';
 
 import {ActivityIndicator, Alert, RefreshControl, ScrollView, StyleSheet, useWindowDimensions} from 'react-native';
-import RenderHTML from 'react-native-render-html';
 import {useNavigation} from '@react-navigation/native';
 
-import {Heading, Text, View, VStack} from 'native-base';
+import {Heading, Text, View} from 'native-base';
 
 import {parseISO} from 'date-fns';
 
-import {
-  AvalancheDangerForecast,
-  AvalancheForecastZone,
-  AvalancheForecastZoneSummary,
-  DangerLevel,
-  ElevationBandNames,
-  ForecastPeriod,
-  Product,
-} from 'types/nationalAvalancheCenter';
-import {AvalancheDangerTable} from './AvalancheDangerTable';
-import {AvalancheDangerIcon} from './AvalancheDangerIcon';
-import {AvalancheProblemCard} from './AvalancheProblemCard';
+import {AvalancheForecastZone, AvalancheForecastZoneSummary} from 'types/nationalAvalancheCenter';
 import {Tab, TabControl} from 'components/TabControl';
 import {useAvalancheForecast} from 'hooks/useAvalancheForecast';
 import {useAvalancheCenterMetadata} from 'hooks/useAvalancheCenterMetadata';
 import {useRefreshByUser} from 'hooks/useRefreshByUser';
+
+import {AvalancheTab} from './AvalancheTab';
 
 export interface AvalancheForecastProps {
   zoneName: string;
@@ -31,61 +21,6 @@ export interface AvalancheForecastProps {
   date: string;
   forecast_zone_id: number;
 }
-
-interface AvalancheTabProps {
-  windowWidth: number;
-  zone: AvalancheForecastZone;
-  forecast: Product;
-}
-
-const AvalancheTab = React.memo(({windowWidth, zone, forecast}: AvalancheTabProps) => {
-  let currentDanger: AvalancheDangerForecast | undefined = forecast.danger.find(item => item.valid_day === ForecastPeriod.Current);
-  if (!currentDanger || !currentDanger.upper) {
-    // sometimes, we get an entry of nulls for today
-    currentDanger = {
-      lower: DangerLevel.None,
-      middle: DangerLevel.None,
-      upper: DangerLevel.None,
-      valid_day: ForecastPeriod.Tomorrow,
-    };
-  }
-  const highestDangerToday: DangerLevel = Math.max(currentDanger.lower, currentDanger.middle, currentDanger.upper);
-
-  let outlookDanger: AvalancheDangerForecast | undefined = forecast.danger.find(item => item.valid_day === ForecastPeriod.Tomorrow);
-  if (!outlookDanger || !outlookDanger.upper) {
-    // sometimes, we get an entry of nulls for tomorrow
-    outlookDanger = {
-      lower: DangerLevel.None,
-      middle: DangerLevel.None,
-      upper: DangerLevel.None,
-      valid_day: ForecastPeriod.Tomorrow,
-    };
-  }
-
-  const elevationBandNames: ElevationBandNames = zone.config.elevation_band_names;
-
-  return (
-    <VStack>
-      <View style={styles.bound}>
-        <AvalancheDangerIcon style={styles.icon} level={highestDangerToday} />
-        <View style={styles.content}>
-          <Text style={styles.title}>THE BOTTOM LINE</Text>
-          <RenderHTML source={{html: forecast.bottom_line}} contentWidth={windowWidth} />
-        </View>
-      </View>
-      <AvalancheDangerTable date={parseISO(forecast.published_time)} current={currentDanger} outlook={outlookDanger} elevation_band_names={elevationBandNames} />
-      <Text style={styles.heading}>Avalanche Problems</Text>
-      {forecast.forecast_avalanche_problems.map((problem, index) => (
-        <AvalancheProblemCard key={`avalanche-problem-${index}`} problem={problem} names={elevationBandNames} />
-      ))}
-      <Text style={styles.heading}>Forecast Discussion</Text>
-      <View style={styles.discussion}>
-        <RenderHTML source={{html: forecast.hazard_discussion}} contentWidth={windowWidth} />
-      </View>
-      <Text style={styles.heading}>Media</Text>
-    </VStack>
-  );
-});
 
 const WeatherTab = () => {
   return <Heading>Weather coming soon</Heading>;
@@ -166,44 +101,3 @@ export const AvalancheForecast: React.FunctionComponent<AvalancheForecastProps> 
     </ScrollView>
   );
 };
-
-const styles = StyleSheet.create({
-  icon: {
-    position: 'absolute',
-    top: -25,
-    left: -25,
-    height: 50,
-  },
-  bound: {
-    margin: 25,
-    borderStyle: 'solid',
-    borderWidth: 1.2,
-    borderColor: 'rgb(200,202,206)',
-    shadowOffset: {width: 1, height: 2},
-    shadowOpacity: 0.8,
-    shadowColor: 'rgb(157,162,165)',
-  },
-  content: {
-    flexDirection: 'column',
-    paddingTop: 15,
-    paddingLeft: 15,
-    paddingBottom: 0,
-    paddingRight: 5,
-  },
-  title: {
-    fontWeight: 'bold',
-  },
-  heading: {
-    textTransform: 'uppercase',
-    fontWeight: 'bold',
-    paddingBottom: 10,
-    marginHorizontal: 10,
-  },
-  discussion: {
-    flexDirection: 'column',
-    paddingTop: 0,
-    paddingLeft: 10,
-    paddingBottom: 0,
-    paddingRight: 10,
-  },
-});
